Drop plugins/cache/base
