+++ tests/sanity/ignore-2.11.txt
@@ -1,4 +1,3 @@
 plugins/action/net_base.py action-plugin-docs # base class for other net_* action plugins which have a matching module
 plugins/action/netconf.py action-plugin-docs # base class for deprecated network platform modules using `connection: local`
 plugins/action/network.py action-plugin-docs # base class for network action plugins
-plugins/cache/base.py ansible-doc!skip # base class for network cache plugins
