+++ tests/sanity/ignore-2.9.txt
@@ -28,4 +28,3 @@ plugins/modules/net_vlan.py validate-modules:deprecation-mismatch
 plugins/modules/net_vlan.py validate-modules:invalid-documentation
 plugins/modules/net_vrf.py validate-modules:deprecation-mismatch
 plugins/modules/net_vrf.py validate-modules:invalid-documentation
-plugins/cache/base.py ansible-doc!skip # base class for network cache plugins
